@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=3.0.0-179.0.dev <4.0.0'
 
 dependencies:
   flutter:
@@ -22,11 +22,12 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.5
   flutter_stage:
     path: ../
-  flare_flutter: ^1.8.3
-  intl: "^0.16.0"
+  flare_flutter: ^3.0.2
+  intl: ^0.18.0
+  google_fonts: ^4.0.3
 
 dev_dependencies:
   flutter_test:
@@ -76,3 +77,6 @@ flutter:
   # see https://flutter.dev/custom-fonts/#from-packages
   assets:
     - assets/clock/
+    - assets/cow.png
+    - assets/dragon.png
+    - assets/rhinoceros.png
